Re-org where some of the new classes are found
